Fixed popular section cards
